Start to support .loc[slice] (unsure whether to upstream) (#37)
* Start to support .loc[slice]
This helps provide limited support for using a slice in .loc calls;
which is a pre-requisite to a lot of other interesting operations.
This only provides support for positional indexes and it does so
by translating to iloc. The next step will be to start providing
slice support for labels properly and then work on the time index.

* A more pandas-like interface to ge/le on an index. Not necessarily efficient
because we are not using sparse arrays, but maybe we can figure that out
later.